Intern Project

+++ src/components/Navbar.jsx
@@ -10,13 +10,18 @@ const Navbar = () => {
   const localSignup = localStorage.getItem("signup");
 
   const handleLogout = () => {
-    localStorage.clear();
+    if (!localSignup) {
+      return navigate("/login");
+    } else {
+      localStorage.clear();
     window.location.reload();
+    }
+    
   };
 
   const handleAccount = () => {
     if (!localSignup) {
-      return navigate("/home");
+      return navigate("/");
     } else {
       return navigate("/admin");
     }
@@ -27,34 +32,39 @@ const Navbar = () => {
       <div className="flex justify-between items-center px-[75px] h-[75px]">
         <p className="text-2xl font-bold ">
           <span className="text-3xl text-[var(--primary-color)] font-bold ">
-            <Link to={"/home"}>Lanceme</Link>
+            <Link to={"/"}>Lanceme  </Link>
           </span>
-          Shop
+          Shop.
         </p>
         <div className="flex gap-4 relative items-center ">
-          {localSignup ? (
+          {!localSignup ? (
             <button
               onClick={handleLogout}
-              className=" bg-[var(--primary-color)] text-white p-2  rounded-lg "
+              className=" bg-[var(--primary-color)] text-white p-2  rounded-lg hover:bg-[var(--secondary-color)]"
             >
-              logout
+              Login
             </button>
-          ) : null}
-          {localSignup ? (
+          ) : <button
+          onClick={handleLogout}
+          className=" bg-[var(--primary-color)] text-white p-2  rounded-lg hover:bg-[var(--secondary-color)]"
+        >
+          Logout
+        </button>}
+          
             <button
               onClick={handleAccount}
-              className=" bg-[var(--primary-color)] text-white p-2  rounded-lg "
+              className=" bg-[var(--primary-color)] text-white p-2  rounded-lg hover:bg-[var(--secondary-color)]"
             >
               Admin
             </button>
-          ) : null}
+          
           <AiOutlineShoppingCart
             className="cursor-pointer"
             onClick={() => navigate("/carts")}
             size={40}
           />
           {cart.cartItem.length !== 0 ? (
-            <div className="flex absolute right-[-12px] top-0 justify-center p-[10px] items-center w-[12px] h-[12px] rounded-[50%] text-white  bg-[var(--primary-color)] ">
+            <div className="flex absolute right-[-12px] top-0 justify-center p-[10px] items-center w-[12px] h-[12px] rounded-[50%] text-white  bg-[var(--primary-color)] hover:bg-[var(--secondary-color)]">
               {cart.cartItem.length}
             </div>
           ) : null}
